backend/requirements.txt
# Backend dependencies
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
supabase==2.9.1
openai==1.55.0
python-multipart==0.0.12
pydantic==2.9.2
gunicorn==21.2.0
